backEnd/evidence_layer/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "evidence-layer"
version = "0.1.0"
description = "PDF ingestion pipeline for Basis - transforms documents into structured, retrievable artifacts"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Basis Team" }
]

dependencies = [
    "pdfplumber>=0.10.0",
    "pymupdf>=1.24.0",
    "tiktoken>=0.5.0",
    "rank-bm25>=0.2.2",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.0",
    "pydantic>=2.0.0",
    "jsonlines>=4.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "google-cloud-storage>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
evidence-layer = "evidence_layer.cli.__main__:app"

[tool.hatch.build.targets.wheel]
packages = ["cli", "src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

